backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
requests==2.31.0
googlemaps==4.10.0
firebase-admin==6.1.0
pydantic==1.10.9
python-dotenv==1.0.0
gunicorn==20.1.0
httpx==0.24.1
email_validator
